fix Comment format

+++ internal/stat/window_test.go
@@ -23,8 +23,8 @@ func TestWindowResetBucket(t *testing.T) {
 	}
 	window.ResetBucket(1)
 	assert.Equal(t, len(window.Bucket(1).Points), 0)
-	assert.Equal(t, window.Bucket(0).Points[0], float64(1.0))
-	assert.Equal(t, window.Bucket(2).Points[0], float64(1.0))
+	assert.Equal(t, window.Bucket(0).Points[0], 1.0)
+	assert.Equal(t, window.Bucket(2).Points[0], 1.0)
 }
 
 func TestWindowResetBuckets(t *testing.T) {
@@ -52,7 +52,7 @@ func TestWindowAdd(t *testing.T) {
 	window := NewWindow(3)
 	window.Append(0, 1.0)
 	window.Add(0, 1.0)
-	assert.Equal(t, window.Bucket(0).Points[0], float64(2.0))
+	assert.Equal(t, window.Bucket(0).Points[0], 2.0)
 }
 
 func TestWindowSize(t *testing.T) {
